# contracts/openapi/m-ranked-v1.yaml
openapi: 3.1.0
info:
  title: M-Ranked API
  version: 1.0.0
  description: >-
    Public reads are backed by revision-pinned projections and legacy identifier aliases.
    Administrative status and account commands are authenticated, role-scoped and never cached.
...
paths:
...
  /api/v1/overview:
    get:
      tags: [Query]
      operationId: getOverview
      summary: Page through institution metrics
      description: >-
        Uses an opaque keyset cursor over a globally sorted legacy-card projection. Missing sort
        values stay last in both directions. Results and validators are pinned to the dataset
        revision selected at the beginning of the request.
      security: []
      parameters:
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Period'
        - name: q
          in: query
          description: Case-insensitive institution-name fragment.
          schema:
            type: string
            maxLength: 200
            default: ''
        - name: sort
          in: query
          description: >-
            Legacy sort key. `coverage` and `accounts` apply only to `all`; `posts` and
            `subscribers` apply only to a concrete platform. Unsupported combinations normalize
            to `m_rating` for `all` and `median_reactions` otherwise.
          schema:
            type: string
            enum: [name, median_reactions, m_rating, reactions, views, posts, subscribers, coverage, accounts]
            default: median_reactions
        - name: direction
          in: query
          description: Defaults to ascending for `name`, descending otherwise.
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          description: Opaque cursor returned as `nextCursor` by a previous response.
          schema:
            type: string
            maxLength: 512
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: A revision-pinned page of metrics.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewPage'
        '304':
          description: The supplied validator matches this representation.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/institutions/{legacyId}:
    get:
      tags: [Query]
      operationId: getInstitution
      summary: Resolve a legacy institution id and return its metrics
      security: []
      parameters:
        - $ref: '#/components/parameters/LegacyId'
        - $ref: '#/components/parameters/Platform'
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Institution and revision-pinned metrics.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Institution'
        '304':
          description: The supplied validator matches this representation.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/publications/{legacyId}:
    get:
      tags: [Query]
      operationId: getPublication
      summary: Resolve a legacy publication id and return latest independent counters
      security: []
      parameters:
        - $ref: '#/components/parameters/LegacyId'
        - name: legacyType
          in: query
          description: Namespace of the legacy identifier.
          schema:
            type: string
            enum: [posts, platform_posts]
            default: posts
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Publication identity and its latest revision-pinned counters.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        '304':
          description: The supplied validator matches this representation.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/rating:
    get:
      tags: [Query]
      operationId: getRating
      summary: Read the legacy activity rating for one platform and period
      description: >-
        Returns activity entities plus the top 50 publications from one dataset revision. Telegram
        ranks enabled channels and keeps channels without publications; VK and RUTUBE aggregate
        institutions that have publications. Metrics come from each publication's coherent latest
        non-synthetic cumulative snapshot. Entities are intentionally bounded by `entityLimit`;
        `entitiesTruncated` reports when this target API cannot return the legacy page's unbounded
        tail. Unknown period values accepted by the browser route fall back to `1d`; omission is
        `30d`. This endpoint is separate from official rating observations and derived rating runs.
      security: []
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [telegram, vk, rutube]
            default: telegram
        - name: period
          in: query
          schema:
            $ref: '#/components/schemas/PeriodValue'
          description: Omission selects 30d; the server retains the legacy invalid-value fallback to 1d.
        - name: channel_sort
          in: query
          schema:
            type: string
            enum: [average, total, engagement, views, subscribers]
            default: engagement
        - name: channel_direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: post_sort
          in: query
          schema:
            type: string
            enum: [reactions, subscriber_share, view_share, views, comments, shares, interactions]
            default: view_share
        - name: post_direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: entityLimit
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 200
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: A bounded, revision-pinned activity rating and top-publication result.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        '304':
          description: The supplied validator matches this representation.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/compare:
    get:
      tags: [Query]
      operationId: getComparison
      summary: Read one fixed-cohort comparison
      description: >-
        Executes one set-based query over the ready comparison projection. The response is bounded
        by a fixed horizon and at most 50 series. For Telegram, repeating `channels` selects legacy
        channel IDs; for VK, MAX, and RUTUBE, repeating `institutions` selects legacy institution
        IDs. Only the parameter matching the selected platform is read; the other legacy namespace
        is ignored for browser compatibility. Duplicate relevant IDs are de-duplicated in first
        occurrence order, and unmapped IDs are ignored without replacing them. When the
        platform-specific selection parameter is absent, `institutionLimit` bounds the
        deterministic default slice. Invalid or excess relevant values fail as an invalid request
        and are never replaced by another series. Every series also includes a companion
        engagement curve calculated as a median of per-publication percentages (never a ratio of
        aggregate medians). Telegram uses reactions/views and starts that curve at hour 1; VK and
        RUTUBE use the available reactions, comments, and shares from the same source observation
        divided by views. Metric and engagement cohorts are independently fixed by their valid
        start/end observations, and intermediate hours only carry earlier valid observations
        forward; future observations are never carried backward.
      security: []
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [telegram, vk, max, rutube]
            default: telegram
        - name: horizonHours
          in: query
          schema:
            type: integer
            enum: [24, 48, 72, 168, 336]
            default: 72
        - name: includePartial
          in: query
          schema:
            type: boolean
            default: false
        - name: metric
          in: query
          schema:
            type: string
            enum: [views, reactions, comments, shares]
            default: reactions
        - name: aggregation
          in: query
          schema:
            type: string
            enum: [sum, median]
            default: median
        - name: institutionLimit
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50
            default: 25
          description: Bounds the default slice when the platform-specific selection is absent.
        - name: channels
          in: query
          description: Ordered legacy channel IDs for Telegram; ignored for every other platform.
          style: form
          explode: true
          schema:
            type: array
            minItems: 1
            maxItems: 50
            items:
              type: integer
              format: int64
              minimum: 1
        - name: institutions
          in: query
          description: Ordered legacy institution IDs for VK, MAX, or RUTUBE; ignored for Telegram.
          style: form
          explode: true
          schema:
            type: array
            minItems: 1
            maxItems: 50
            items:
              type: integer
              format: int64
              minimum: 1
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: One ready, revision-pinned fixed cohort.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comparison'
        '304':
          description: The supplied validator matches this representation.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/accounts/{legacyId}:
    get:
      tags: [Query]
      operationId: getAccount
      summary: Resolve a legacy channel or platform-account id
      description: >-
        Returns canonical catalog metadata and a small aggregate from `publication_latest`.
        Subscriber history is intentionally absent because no account-latest projection exists.
      security: []
      parameters:
        - $ref: '#/components/parameters/LegacyId'
        - name: legacyType
          in: query
          schema:
            type: string
            enum: [channels, platform_accounts]
            default: platform_accounts
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Alias-backed account metadata and projection summary.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '304':
          description: The supplied validator matches this representation.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/PublicCache'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        Transitional stateless admin authentication. Store only bcrypt hashes in runtime
        configuration and expose this surface exclusively behind TLS. OIDC remains a later slice.
  parameters:
    Platform:
      name: platform
      in: query
      schema:
        type: string
        enum: [all, telegram, vk, max, rutube]
        default: all
    Period:
      name: period
      in: query
      schema:
        type: string
        enum: [3h, 1d, 7d, 30d]
        default: 1d
    LegacyId:
      name: legacyId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: Current or weak ETag; wildcard matching is supported.
      schema:
        type: string
  headers:
    ETag:
      description: >-
        Validator derived from the same API namespace, authoritative PostgreSQL dataset revision,
        and normalized-query fingerprint used by the public response cache.
      schema:
        type: string
    PublicCache:
      description: Public response cache policy.
      schema:
        type: string
        const: max-age=30, must-revalidate, public
    NoStore:
      description: Response must not be stored.
      schema:
        type: string
        const: no-store
  responses:
    InvalidRequest:
      description: A query, path parameter, or cursor is invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No alias-backed resource exists at this dataset revision.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: An unexpected failure occurred before the response was committed.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    OverviewPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, datasetRevision, asOf]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OverviewRow'
        nextCursor:
          type: [string, 'null']
        datasetRevision:
          type: integer
          format: int64
          minimum: 0
        asOf:
          type: string
          format: date-time
    Institution:
      type: object
      additionalProperties: false
      required:
        - institutionId
        - legacyId
        - canonicalName
        - shortName
        - platform
        - period
        - metrics
        - datasetRevision
        - asOf
      properties:
        institutionId:
          type: string
          format: uuid
        legacyId:
          type: integer
          format: int64
        canonicalName:
          type: string
        shortName:
          type: [string, 'null']
        platform:
          $ref: '#/components/schemas/PlatformValue'
        period:
          $ref: '#/components/schemas/PeriodValue'
        metrics:
          $ref: '#/components/schemas/Metrics'
        datasetRevision:
          type: integer
          format: int64
          minimum: 0
        asOf:
          type: string
          format: date-time
...
    Publication:
      type: object
      additionalProperties: false
      required:
        - publicationId
        - legacyId
        - legacyType
        - institutionId
        - platform
        - publishedAt
        - publicationType
        - deletedAt
        - views
        - reactions
        - comments
        - shares
        - quality
        - intervalUncertain
        - synthetic
        - historyCompleteness
        - datasetRevision
        - asOf
      properties:
        publicationId:
          type: string
          format: uuid
        legacyId:
          type: integer
          format: int64
        legacyType:
          type: string
          enum: [posts, platform_posts]
        institutionId:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/PlatformValue'
        publishedAt:
          type: string
          format: date-time
        publicationType:
          type: string
        deletedAt:
          type: [string, 'null']
          format: date-time
        views:
          $ref: '#/components/schemas/CounterMetric'
        reactions:
          $ref: '#/components/schemas/CounterMetric'
        comments:
          $ref: '#/components/schemas/CounterMetric'
        shares:
          $ref: '#/components/schemas/CounterMetric'
        quality:
          type: [string, 'null']
        intervalUncertain:
          type: boolean
        synthetic:
          type: boolean
        historyCompleteness:
          type: [string, 'null']
        datasetRevision:
          type: integer
          format: int64
          minimum: 0
        asOf:
          type: string
          format: date-time
...
    PeriodValue:
      type: string
      enum: [3h, 1d, 7d, 30d]
...